+++ src/test/java/de/dm/intellij/liferay/index/RenderCommandIndexTest.java
@@ -6,6 +6,7 @@ import com.intellij.openapi.projectRoots.impl.JavaAwareProjectJdkTableImpl;
 import com.intellij.openapi.roots.ContentEntry;
 import com.intellij.openapi.roots.LanguageLevelModuleExtension;
 import com.intellij.openapi.roots.ModifiableRootModel;
+import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess;
 import com.intellij.pom.java.LanguageLevel;
 import com.intellij.psi.search.GlobalSearchScope;
@@ -36,9 +37,9 @@ public class RenderCommandIndexTest extends LightCodeInsightFixtureTestCase {
             model.setSdk(jdk);
 
             final String testDataPath = PathUtil.toSystemIndependentName(new File(TEST_DATA_PATH).getAbsolutePath());
-            VfsRootAccess.allowRootAccess( testDataPath );
+            VfsRootAccess.allowRootAccess( Disposer.newDisposable(), testDataPath );
 
-            PsiTestUtil.addLibrary(module, model, "OSGi", testDataPath, "osgi.jar");
+            PsiTestUtil.addLibrary(model, "OSGi", testDataPath, "osgi.jar");
         }
     };