+++ src/test/java/de/dm/intellij/liferay/language/velocity/LiferayVtlVariableProviderTest.java
@@ -6,6 +6,7 @@ import com.intellij.openapi.projectRoots.Sdk;
 import com.intellij.openapi.roots.ContentEntry;
 import com.intellij.openapi.roots.LanguageLevelModuleExtension;
 import com.intellij.openapi.roots.ModifiableRootModel;
+import com.intellij.openapi.util.Disposer;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess;
 import com.intellij.pom.java.LanguageLevel;
@@ -42,11 +43,11 @@ public class LiferayVtlVariableProviderTest extends LightCodeInsightFixtureTestC
 
             URL resource = LiferayFtlVariableProviderTest.class.getResource("/com/liferay/vtl/VM_liferay_70.vm");
             String resourcePath = PathUtil.toSystemIndependentName(new File(resource.getFile()).getAbsolutePath());
-            VfsRootAccess.allowRootAccess( resourcePath );
+            VfsRootAccess.allowRootAccess( Disposer.newDisposable(), resourcePath );
 
             resource = LiferayFtlVariableProviderTest.class.getResource("/com/liferay/tld");
             resourcePath = PathUtil.toSystemIndependentName(new File(resource.getFile()).getAbsolutePath());
-            VfsRootAccess.allowRootAccess( resourcePath );
+            VfsRootAccess.allowRootAccess( Disposer.newDisposable(), resourcePath );
 
         }