+++ testdata/de/dm/intellij/liferay/language/osgi/LiferayOsgiImplicitUsageProviderTest/MyComponent.java
@@ -1,6 +1,6 @@
 import org.osgi.service.component.annotations.Reference;
 
-public class <warning descr="Class 'MyComponent' is never used">MyComponent</warning> {
+public class MyComponent {
 
     @Reference
     private org.osgi.service.component.annotations.ConfigurationPolicy <warning descr="Private field 'foo' is assigned but never accessed">foo</warning>;